+++ src/Interop/LibHeif/NativeMethods/LibHeifNative.Image.cs
@@ -111,5 +111,15 @@ namespace LibHeifSharp.Interop
                                                                            int firstWarningIndex,
                                                                            heif_error* outputBuffer,
                                                                            int maxOutputBufferEntries);
+
+        [DllImport(DllName, CallingConvention = DllCallingConvention)]
+        [return: MarshalAs(UnmanagedType.Bool)]
+        internal static extern bool heif_image_has_content_light_level(SafeHeifImage image);
+
+        [DllImport(DllName, CallingConvention = DllCallingConvention)]
+        internal static extern void heif_image_get_content_light_level(SafeHeifImage image, out heif_content_light_level value);
+
+        [DllImport(DllName, CallingConvention = DllCallingConvention)]
+        internal static extern unsafe void heif_image_set_content_light_level(SafeHeifImage image, heif_content_light_level* value);
     }
 }